+++ src/utils/getContractAddress.js
@@ -1,0 +1,6 @@
+import ethUtil from 'ethereumjs-util'
+
+const getContractAddress = (account, nonce) =>
+  ethUtil.bufferToHex(ethUtil.generateAddress(account, nonce))
+
+export default getContractAddress